# perf: defer the allocation of matchValue.