`matchValue` isn't needed on the fast path, so don't allocate it until it's actually needed. Additionally, use a Span-based conversion to UTF-8 bytes to avoid the creation of a temporary char array.

```csharp
[Benchmark]
public int CountTokens() => m_encoder.CountTokens(Text);

[Benchmark]
public IReadOnlyCollection<int> Encode() => m_encoder.Encode(Text);
```

```
BenchmarkDotNet v0.14.0, Windows 10 (10.0.19045.5073/22H2/2022Update)
Intel Core i7-10875H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.100-rc.2.24474.11
  [Host] : .NET 9.0.0 (9.0.24.47305), X64 RyuJIT AVX2
  2.1.0  : .NET 9.0.0 (9.0.24.47305), X64 RyuJIT AVX2
  PR     : .NET 9.0.0 (9.0.24.47305), X64 RyuJIT AVX2
```

| Method      | Job   | NuGetReferences       | Mean      | Error     | StdDev    | Median   | Gen0   | Allocated |
|------------ |------ |---------------------- |----------:|----------:|----------:|---------:|-------:|----------:|
| CountTokens | 2.1.0 | Tiktoken 2.1.0        | 10.893 us | 1.6238 us | 2.3802 us | 9.392 us | 0.6104 |   5.08 KB |
| CountTokens | PR    | This PR |  8.482 us | 0.2161 us | 0.3168 us | 8.421 us | 0.2899 |   2.49 KB |
| | | | | | | | | |
| Encode      | 2.1.0 | Tiktoken 2.1.0        |  9.396 us | 0.2398 us | 0.3362 us | 9.269 us | 0.9155 |   7.56 KB |
| Encode      | PR    | This PR |  8.802 us | 0.1064 us | 0.1592 us | 8.790 us | 0.5951 |   4.98 KB |


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
	- Enhanced text encoding and tokenization processes for improved performance.
	- Introduced a new method for efficient conversion of character spans to byte spans.

- **Bug Fixes**
	- Optimized memory usage during token processing, reducing unnecessary allocations.

- **Documentation**
	- Updated internal documentation to reflect changes in tokenization and encoding methods.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->