@HavenDV 2024-11-09
It's just a bot, probably using gpt4o-mini or worse models, so don't expect any insight into such matters. But sometimes it works