@bgrainger 2024-11-09
I've heard of it before, but never seen it enabled on a project. I know to take its suggestions with a (large) grain of salt, but couldn't resist defending myself against the incorrect AI 😂

(If anything it seems worthwhile to note the objections to its review for future maintainers/contributors who are looking into the history of the code.)